May 19/web_scraping_and_summarizer/requirements.txt
selenium==4.19.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
google-generativeai==0.5.2
python-dotenv==1.0.1
